restructured rust code org
