+++ src/num/cond_entropy.rs
