+++ src/num/haversine.rs
