+++ src/num/knn.rs
