+++ src/stats/sample.rs
