+++ src/str2/aho_corasick.rs
