+++ src/str2/consts.rs
