+++ src/str2/fuzz.rs
