+++ src/str2/hamming.rs
