+++ src/str2/inflections.rs
