+++ src/str2/osa.rs
