+++ src/str2/snowball/algorithms/english_stemmer.rs
@@ -5,8 +5,8 @@
 #![allow(unused_mut)]
 #![allow(unused_parens)]
 #![allow(unused_variables)]
-use crate::str_ext::snowball::Among;
-use crate::str_ext::snowball::SnowballEnv;
+use crate::str2::snowball::Among;
+use crate::str2::snowball::SnowballEnv;
 
 static A_0: &'static [Among<Context>; 3] = &[
     Among("arsen", -1, -1, None),
