+++ src/str2/snowball/algorithms/mod.rs
